docs : feature 디렉토리 구조에 맞게 코드 리펙토링
