+++ src/main/java/DevBackEnd/ForLong/features/user/dto/FindUserDTO.java
@@ -1,4 +1,4 @@
-package DevBackEnd.ForLong.features.login.dto;
+package DevBackEnd.ForLong.features.user.dto;
 
 import DevBackEnd.ForLong.core.entity.Pet;
 import DevBackEnd.ForLong.core.entity.User;
